Move GetLocalHosts to usecase
